[Feat]: Studio header 구성, upload 연결
 - StudioHeader.jsx: useRef, props 사용해서 upload 컴포넌트 연결
 - StudioHeader.scss: z-index 수정, position 설정
 - Upload.jsx: upload file picker 컴포넌트 분리, upload dialog open, close 구현
 - UploadFilePicker.jsx: upload file picker 컴포넌트 분리
