+++ src/App.css
@@ -117,6 +117,7 @@ wt-iron-overlay-backdrop {
   opacity: 0;
   transition: opacity 0.2s;
   z-index: 2002;
+  display: none;
 }
 
 wt-icon-button {
@@ -164,6 +165,7 @@ wt-upload-file-picker {
 
 wt-upload-dialog {
   display: flex;
+  display: none;
   flex-direction: column;
   --paper-dialog-scrollable-padding: 0;
   --header-title-row-height: 64px;
