+++ src/components/Header/StudioHeader.scss
@@ -10,10 +10,9 @@ wt-st-header {
     background-color: var(--studio-bg);
     color: var(--text-primary);
     box-shadow: 0px 1px 4px 1px var(--wt-overlay-drop-shadow-app-header);
-}
+    z-index: 100;
+    position: fixed;
 
-wt-st-header.wt-entity-page {
-    z-index: 2;
 }
 
 header.wt-st-header {
